Move test status badge to components directory
