+++ apps/spruce/src/pages/task/taskTabs/testsTable/getColumnsTemplate.tsx
@@ -1,9 +1,9 @@
 import { WordBreak } from "components/styles";
+import TestStatusBadge from "components/TestStatusBadge";
 import { testStatusesFilterTreeData } from "constants/test";
 import { TestSortCategory, TaskQuery } from "gql/generated/types";
 import { string } from "utils";
 import { LogsColumn } from "./LogsColumn";
-import TestStatusBadge from "./TestStatusBadge";
 
 const { msToDuration } = string;
